feat: restructure server imports
