+++ packages/sdk/src/codegen/templates/typescript/webhooks.template.ts
@@ -1,6 +1,6 @@
 //language=handlebars
 export const template = `
-import { WebhookConfiguration } from '@wundergraph/sdk';
+import type { WebhookConfiguration } from '@wundergraph/sdk';
 
 export type WebhooksConfig = {
 {{#each webhooks}}
