+++ packages/sdk/src/index.ts
@@ -1,4 +1,4 @@
-export type { ResolvedWunderGraphConfig, HooksConfiguration, OperationHooksConfiguration } from './configure';
+export type { ResolvedWunderGraphConfig } from './configure';
 export type {
 	BaseOperationConfiguration,
 	QueryConfiguration,
@@ -12,44 +12,8 @@ export type {
 	CustomizeMutation,
 	CustomizeSubscription,
 } from './configure/operations';
+export { configureWunderGraphServer } from './server';
 export type { GraphQLOperation } from './graphql/operations';
-export type {
-	WunderGraphRequest,
-	WunderGraphResponse,
-	WsTransportOnConnectionInitResponse,
-	WunderGraphUser,
-	BaseRequestContext,
-	AuthenticationRequestContext,
-	AuthenticationResponse,
-	ClientRequest,
-	WunderGraphHooksAndServerConfig,
-	WunderGraphServerConfig,
-	ClientRequestHeaders,
-	AuthenticationHookRequest,
-} from './middleware/types';
-export type {
-	SubscriptionRequestOptions,
-	Response,
-	ResponseOK,
-	RequestOptions,
-	MutateRequestOptions,
-	CachedResponse,
-	UploadResponse,
-	ClientConfig,
-	RequiresAuthentication,
-	Error,
-	Headers,
-	Lazy,
-	None,
-	Loading,
-	Refetch,
-	UserListener,
-	Aborted,
-	FetchConfig,
-} from './middleware/client';
-export type { OperationArgsWithInput, InternalClient } from './middleware/internal-client';
-export type { GraphQLServerConfig } from './middleware/plugins/graphql';
-export type { FastifyLoggerInstance } from 'fastify';
 export type { Template, TemplateOutputFile } from './codegen/index';
 export type {
 	WebhookHttpEvent,
@@ -62,18 +26,14 @@ export type {
 	WebhookConfiguration,
 } from './webhooks/types';
 export type { WebhookVerifierConfiguration } from './webhooks/verifiers';
-
 export { default as templates } from './codegen/templates';
 export { introspect, createMockApi, Api } from './definition';
 export { configureWunderGraphApplication } from './configure';
 export { configureWunderGraphOperations, enableAuth, enableCaching, disableAuth } from './configure/operations';
-export { configureWunderGraphServer } from './middleware/server';
-
 export { default as cors } from './cors';
 export { authProviders } from './configure/authentication';
 export type { LoggerLevel } from './configure/options';
 export { WgEnv } from './configure/options';
-
 export {
 	BaseTypeScriptDataModel,
 	formatTypeScript,
