+++ testapps/default/.wundergraph/wundergraph.config.ts
@@ -110,12 +110,6 @@ configureWunderGraphApplication({
 				  ]
 				: ['http://localhost:3000'],
 	},
-	authentication: {
-		cookieBased: {
-			providers: [authProviders.demo()],
-			authorizedRedirectUris: ['http://localhost:3000'],
-		},
-	},
 	security: {
 		enableGraphQLEndpoint: true,
 	},
